[python] Re-enabled python testing
Found bug in how .dot method is exposed in the bindings,
disabled for now and opened bug at https://github.com/robotology/idyntree/issues/105
