+++ bindings/python/CMakeLists.txt
@@ -35,3 +35,8 @@ string(STRIP ${PYTHON_INSTDIR} PYTHON_INSTDIR)
 if(WIN32)
     set_target_properties(_${target_name} PROPERTIES SUFFIX ".pyd")
 endif(WIN32)
+
+# if compile tests execute also python tests
+if(IDYNTREE_COMPILE_TESTS)
+    add_subdirectory(tests)
+endif()
