+++ bindings/python/tests/CMakeLists.txt
@@ -1,0 +1,7 @@
+
+add_custom_target(copy_python_tests_in_build ALL)
+add_custom_command(TARGET copy_python_tests_in_build PRE_BUILD
+                   COMMAND ${CMAKE_COMMAND} -E copy_directory
+                   ${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_CURRENT_BINARY_DIR})
+
+add_test(NAME PythonGeometryTests COMMAND ${PYTHON_EXECUTABLE} geometry.py)
